ajax try

+++ src/main/java/edu/du/teamportpoilo_pyl/controller/BoardController.java
@@ -4,9 +4,15 @@ import edu.du.teamportpoilo_pyl.dto.BoardDto;
 import edu.du.teamportpoilo_pyl.dto.CommentDto;
 import edu.du.teamportpoilo_pyl.service.BoardService;
 import edu.du.teamportpoilo_pyl.service.CommentService;
+import jakarta.servlet.http.HttpSession;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.HashMap;
+import java.util.Map;
 
 @Controller
 @RequestMapping("/")
@@ -60,9 +66,20 @@ public class BoardController {
     }
 
     @PostMapping("/board/{id}/comment/delete")
-    public String deleteComment(@PathVariable Long id, @ModelAttribute CommentDto commentDto) {
+    public ResponseEntity<Map<String, Object>> deleteComment(@PathVariable Long id,
+                                                             @ModelAttribute CommentDto commentDto,
+                                                             HttpSession session) {
+        Map<String, Object> response = new HashMap<>();
+        String nameSession = (String) session.getAttribute("name");
+
+        if (!commentDto.getAuthor().equals(nameSession)) {
+            response.put("success", false);
+            response.put("message", "게시글 삭제 권한이 없습니다.");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(response);
+        }
+
         commentService.deleteComment(commentDto.getId());
-        return "redirect:/board/" + id;
+        return ResponseEntity.ok(response);
     }
 
     @GetMapping("/board/write")
@@ -83,7 +100,7 @@ public class BoardController {
     }
 
     @PostMapping("board/{id}/delete")
-    public String deleteBoard(@PathVariable Long id) {
+    public String deleteBoard(@PathVariable Long id, HttpSession session) {
         boardService.deleteBoard(id);
         return "redirect:/list";
     }
